rift: Blobwatch improvements
Make x/y position of a blob be floating point, so we can track
sub-pixel positions.

Extend the LED pattern to an int, so we have room to store the
device ID of the LED in the (8..15) bits.

Expand the number of extents and blobs we allow, and permit
smaller 2 pixel wide extents, as the CV1 has many more LEDs
involved, and they are smaller.

Get ready to track a different pixel threshold for CV1, as
it excludes background noise a lot better, and has faint LEDs
that go to lower pixel values.

+++ src/drv_oculus_rift/rift-sensor-flicker.c
@@ -32,7 +32,7 @@ static int hamming_distance(uint16_t a, uint16_t b)
 #endif
 
 static int pattern_find_id(rift_led *leds, int num_patterns,
-			   uint16_t pattern, int8_t *id)
+			   uint16_t pattern, int *id)
 {
 	int i;
